Changed transaction amount type to string, validate route according to latest swagger spec

+++ transaction/transaction.go
@@ -3,7 +3,7 @@ package transaction
 type Transaction struct {
 	EntityId            string
 	Amount              uint64
-	MinorUnits          uint64
+	MinorUnits          int
 	CurrencyCode        CurrencyCode
 	CustomerCountryCode CountryCodeIso31661Alpha2
 }
